rename player movement scripts and add parent class
