+++ Assets/Players/scripts/PlayerMovementScript.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+using System.Collections;
+
+abstract public class PlayerMovementScript : MonoBehaviour {
+
+	void test() {
+	
+	}
+
+}
