+++ Assets/Players/scripts/PlayerMovementWASDScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c9e0b0c54e99e7641b2875c2df6bb9ea
+timeCreated: 1474343626
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
